fix: sigwinch on windows
